+++ packages/client/src/index.ts
@@ -69,6 +69,8 @@ export type {
   MagicEdenBitcoinEvents,
 } from './connectors/magicEden.js'
 
+// client types
+export type { BtcAccount } from './types/account.js'
 export type { State } from './types/state.js'
 export { type Config, createConfig } from './factories/createConfig.js'
 export type { CreateConnectorFn, Connector } from './types/connector.js'