+++ packages/core/src/actions/getBalance.ts
@@ -1,8 +1,8 @@
-import type { Transport } from '../factories/createTransport.js'
 import type { UTXOSchema } from '../transports/types.js'
 import type { Account } from '../types/account.js'
 import type { Chain } from '../types/chain.js'
 import type { Client } from '../types/client.js'
+import type { Transport } from '../types/transport.js'
 
 export type GetBalanceParameters = {
   /** The address of the account. */
@@ -25,5 +25,6 @@ export async function getBalance<
     },
     { dedupe: true }
   )
+
   return data
 }