+++ packages/core/src/actions/getBlockCount.ts
@@ -1,8 +1,8 @@
-import type { Transport } from '../factories/createTransport.js'
 import type { UTXOSchema } from '../transports/types.js'
 import type { Account } from '../types/account.js'
 import type { Chain } from '../types/chain.js'
 import type { Client } from '../types/client.js'
+import type { Transport } from '../types/transport.js'
 
 export type GetBlockCountReturnType = number