+++ packages/core/src/actions/getBlockStats.ts
@@ -1,10 +1,10 @@
 import { BlockNotFoundError } from '../errors/block.js'
-import type { Transport } from '../factories/createTransport.js'
 import type { UTXOSchema } from '../transports/types.js'
 import type { Account } from '../types/account.js'
 import type { BlockStats, BlockStatsKeys } from '../types/blockStats.js'
 import type { Chain } from '../types/chain.js'
 import type { Client } from '../types/client.js'
+import type { Transport } from '../types/transport.js'
 
 export type GetBlockStatsParameters = (
   | {