+++ packages/core/src/actions/signPsbt.ts
@@ -1,4 +1,3 @@
-import type { Transport } from '../factories/createTransport.js'
 import type {
   SignPsbtParameters,
   SignPsbtReturnType,
@@ -7,6 +6,7 @@ import type {
 import type { Account } from '../types/account.js'
 import type { Chain } from '../types/chain.js'
 import type { Client } from '../types/client.js'
+import type { Transport } from '../types/transport.js'
 
 export async function signPsbt<
   C extends Chain | undefined,