+++ packages/core/src/clients/decorators/wallet.ts
@@ -4,15 +4,14 @@ import {
   sendUTXOTransaction,
 } from '../../actions/sendUTXOTransaction.js'
 import { signPsbt } from '../../actions/signPsbt.js'
-import type { Transport } from '../../factories/createTransport.js'
-import type { Account } from '../../types/account.js'
-import type { Client } from '../../types/client.js'
-
 import type {
   SignPsbtParameters,
   SignPsbtReturnType,
 } from '../../transports/types.js'
+import type { Account } from '../../types/account.js'
 import type { Chain } from '../../types/chain.js'
+import type { Client } from '../../types/client.js'
+import type { Transport } from '../../types/transport.js'
 
 export type WalletActions = {
   /**