+++ packages/core/src/transports/blockcypher/blockcypher.ts
@@ -1,0 +1,18 @@
+import type { HttpTransportConfig } from '../http.js'
+import type { UTXOMethod } from '../types.js'
+import { utxo } from '../utxo.js'
+import { blockcypherMethods } from './methods.js'
+
+type BlockcypherConfig = {
+  baseUrl?: string
+  apiKey?: string
+} & HttpTransportConfig
+
+export const blockcypher = (config?: BlockcypherConfig) =>
+  utxo(config?.baseUrl || 'https://api.blockcypher.com/v1/btc/main', {
+    key: 'blockcypher',
+    methods: {
+      include: Object.keys(blockcypherMethods) as UTXOMethod[],
+    },
+    ...config,
+  })